### 2021-05-14 #3287 @whedon
[PRE REVIEW]: pyREX: Workflow-Orientated Python Package for Replica Exchange Molecular Dynamics
**Submitting author:** @ArtVoro (<a href="http://orcid.org/0000-0002-5804-5151">Arthur Voronin</a>)
**Repository:** <a href="https://github.com/KIT-MBS/pyREX" target ="_blank">https://github.com/KIT-MBS/pyREX</a>
**Version:** 1.0
