### 2021-05-14 #3287 @jgostick
@ArtVoro, we have a checklist of things that the reviewers will look for, and I can already tell that your documentation will raise some flags, since there isn't any.  :-)  I can see that you have a ``doc`` folder so I'm sure you're building it and hosting it somewhere, but there is no link to this in your readme or anywhere on the github page.  Could you maybe address this first, so the reviewers and I can start digging in to this package?
