### 2021-05-14 #3287 @jgostick
@ArtVovo, I should also add that your package name is essentially invisible to google since people have run molecular dynamics simulations on actual pyrex brand glass, and that is the only hits I'm getting.  I'm sure you are aware of this, but maybe consider a name change before JOSS chisels it in stone?
