### 2021-05-14 #3287 @ArtVoro
> @ArtVoro, we have a checklist of things that the reviewers will look for, and I can already tell that your documentation will raise some flags, since there isn't any. :-) I can see that you have a `doc` folder so I'm sure you're building it and hosting it somewhere, but there is no link to this in your readme or anywhere on the github page. Could you maybe address this first, so the reviewers and I can start digging in to this package?

@jgostick i will try to fix it asap. until 2 days ago the documentation was just the doc strings of the py files.

> @ArtVovo, I should also add that your package name is essentially invisible to google since people have run molecular dynamics simulations on actual pyrex brand glass, and that is the only hits I'm getting. I'm sure you are aware of this, but maybe consider a name change before JOSS chisels it in stone?

@jgostick i will discuss some alternative names with my colleagues.
