### 2021-05-14 #3287 @jgostick
Good stuff.  RE: name, ``pyrexmd`` might be enough?
