### 2021-05-14 #3287 @ArtVoro
funny i was thinking the same. typically people in this field use REMD for Replica Exchange Molecular Dynamics but i really love REX so i tryed to stick to that.
