### 2021-05-16 #3287 @ArtVoro
@jgostick

1) i renamed the package to `pyrexMD` and fixed all files where i call it
2) i managed to build some api doc with sphinx, will add some stuff later to pep it up a bit
3) i uploaded the doc to github pages and linked it to README
