### 2021-05-18 #3287 @ArtVoro
@jgostick I added more stuff to docs and reworked the doc tree, should be good now i guess.

does the package renaming matter and i have to resubmit or is everything fine as it is?
