### 2021-06-01 #3325 @whedon
[REVIEW]: pyrexMD: Workflow-Orientated Python Package for Replica Exchange Molecular Dynamics
**Submitting author:** @ArtVoro (<a href="http://orcid.org/0000-0002-5804-5151">Arthur Voronin</a>)
**Repository:** <a href="https://github.com/KIT-MBS/pyREX" target ="_blank">https://github.com/KIT-MBS/pyREX</a>
**Version:** v1.0

I believe that this check mark is what has been causing the delay for  both me and @rosecers. Unfortunately, when this paper was first submitted, the package was not well-documented and was hard to understand. We have asked for some clarification and code clean up. Unfortunately, I feel that even with the improvements this work does not meet the requirements of JOSS for substantial scholarly effort and I can't recommend this package for publication in JOSS at this time.

JOSS defines the following for substantial scholarly effort:

As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:

- Age of software (is this a well-established software project) / length of commit history.
- Number of commits.
- Number of authors.
> This package has only one author.

- Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
> While this submission appears to fit this criteria, upon closer inspection this is questionable. Many of the functions in `misc.py`, for example do not add any functionality that are not in Python already (for example

> This package has many more lines of code than are needed to achieve its objective.  Many of the functions are what I would call light wrappers around other libraries and do not add much functionality themselves. Many functions (particularly in `misc.py` reimplement functionality that are already available in Python or have statements that are unnecessary.

- Whether the software has already been cited in academic papers.
> The software has not been cited in academic papers

- Whether the software is sufficiently useful that it is likely to be cited by your peer group.
> I am not convinced of this point because the software is mostly thin wrappers around existing code.

I unfortunately cannot recommend this software package for publication in JOSS at this time. I think it would be more appropriate for the author to obtain a DOI through something like Zenodo and resubmit for publication if the package proves sufficiently useful.
