### 2021-06-15 #3325 @janash
Hi @whedon - I've gone over the COI and general checks and installed the dependencies. Planning to closely review the software later this week and this weekend!
