### 2021-06-18 #3325 @rosecers
> Conflict of interest
> - [X] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> Code of Conduct
> - [X]  I confirm that I read and will adhere to the JOSS code of conduct.
> General checks
> - [X] Repository: Is the source code for this software available at the repository url?
> - [X] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
> - [X] Contribution and authorship: Has the submitting author (@ArtVoro) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> - [ ] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

I'm not convinced of the scholarly effort. My largest concern comes from the citability and need within the community -- there are no publications that mention "pyrexMD", and the publications of the two authors using REMD (3 as of 2020) have not yet been cited*. To this end, the authors need to comment on two things: the breadth of scope of this software (it appears quite narrow), and whether similar workflows would be supported by pre-existing software (particularly [SSAGES](https://ssagesproject.github.io/docs/Engines.html)).
*All citation metrics checked via searching WebOfKnowledge and GoogleScholar.

> Functionality
> - [X] Installation: Does installation proceed as outlined in the documentation?
> - [X]  Functionality: Have the functional claims of the software been confirmed?

The results contained in the paper summary are reproducible in the examples. Beyond that, no functional claims have been made.

> - [X] Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

> Documentation
> - [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The authors have an explicit statement of need in their write-up; however, mentioned above, they have not addressed the breadth of the scope of this package and the currently available software packages intended to this end.

> - [ ]  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The requirements for pyrexMD need to be more prominently featured. For example, some packages require specific python versions, which may not be clear upon installation. `pytest` and `autopep8`should be added as a requirement for testing and examples purposes.

> - [ ] Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The authors include examples, however, these are not well-documented and require more explanation / commenting. Furthermore, there need to be embedded examples in the documentation -- if there is a user considering using this package, they may want to see a demonstration of the package before downloading and without searching through the API.

> - [X] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There is API documentation -- it is sufficient, but could greatly benefit from further formatting and editing.

> - [X] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

The tests are sufficient, however, they currently raise several warnings, which should be remedied. As for coverage, no coverage test or CI is implemented to check on the functionality.

> - [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No.

> Software paper
> - [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary is minimal -- the authors should expand considerably on the REX methodology and the workflows supported by pyrexMD.

> - [ ] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

There is an explicit statement of need that needs to be amended with respect to the comments in the preceding sections.

> - [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

No.

> - [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The authors are fairly terse in their writing, and need to expand in several places (e.g. the figures need substantially more explanation).

> - [X] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Overall, the examples, use cases, and need for this software need to be expanded upon. The examples should include commentary as to the expected behavior and "behind the scenes" operation of the software, including an introductory paragraph at the top of the example to explain what is being demonstrated. The authors have not identified use cases for this software beyond their own publications -- they should detail in their write-up examples of the sorts analyses possible with this software. Finally, there needs to be some statement as to the novelty of this software -- is it fulfilling a need within the community which is not currently being met?
