### 2021-06-18 #3325 @ArtVoro
thank you for the feedback. i will try to solve most of the things mentioned over this weekend.
