### 2021-06-20 #3325 @janash
> ### Conflict of interest
>
> - [x] I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
>
> - [x] I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
>
> - [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="https://github.com/KIT-MBS/pyREX">repository url</a>?
> - [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> - [x] **Contribution and authorship:** Has the submitting author (@ArtVoro) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> - [ ] **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

This software mainly seems to be a wrapper around three libraries - gromacs, nglview, and mdanalysis. The software has not yet been cited in academic papers. In its current state, the paper does not appear to meet the definition of substantial scholarly effort.

> ### Functionality
>
> - [ ] **Installation:** Does installation proceed as outlined in the documentation?

I had difficulty installing the software due to a missing dependency with mpi4py on my mac. I used `brew install mpich` which then allowed my installation to proceed. **I recommend adding a CI action ([GitHub Actions](https://docs.github.com/en/actions) or other of your choice) to test installation on supported operating systems.** You should also run your tests using CI and add code coverage.

> - [] **Functionality:** Have the functional claims of the software been confirmed?

I feel that I need expanded example usage documentation to determine this. The main functional claims of the software are in the last paragraph of the statement of need. It might be beneficial to number the examples as tutorials and reference the tutorial numbers in the paper.

> - [x] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

No performance claims made.

>
> ### Documentation
>
> - [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> - [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

There is a list of installation instructions. I did encounter an error with dependencies as described above. I have a few comments about the installation instructions:
- there is no need to do an "editable' `-e` install for most users, unless you expect them to be pulling from the repo for updates.
- I recommend putting the package on the [Python Package Index (PyPI)](https://pypi.org/) so that users can install with pip.

> - [ ] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There is example usage, but the documentation associated with the example usage needs to be expanded. Please use a combination of markdown and code cells in your example notebook to give an explanation of tutorials.

> - [x] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Please move documentation link on the readme to top. There is API documentation that is adequate for using the package. However, it would benefit from additional documentation or better organization.

> - [ ] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

There is a set of tests which can be run with pytest (manual steps), but there are no automated tests.

> - [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No.

> ### Software paper
>
> - [x] **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> - [x] **A statement of need:** Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
> - [ ] **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

No. Please expand this portion.

> - [x] **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

> - [ ] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

**Summary**
The author needs to give more justification for the scholarly effort of this package as suggested by @rosecers . The documentation should be improved and expanded, with more explanation given for the examples. A tutorial describing a full workflow (set-up, running simulation, and analysis) should be provided as well.

I also suggest automated building and testing of your package using GitHub Actions, and putting your package on PyPI for ease of installation.
