### 2021-06-21 #3325 @rosecers
> > ### Functionality
> >
> > * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
>
> I had difficulty installing the software due to a missing dependency with mpi4py on my mac. I used `brew install mpich` which then allowed my installation to proceed. **I recommend adding a CI action ([GitHub Actions](https://docs.github.com/en/actions) or other of your choice) to test installation on supported operating systems.** You should also run your tests using CI and add code coverage.

I'd like to echo @janash's difficulty with installation, as I also had issue with mpi4py on my mac; I concur that the authors should test installation on supported operating systems. Ultimately, I needed to run:

```
cd $INSTALL_DIR
git clone https://github.com/KIT-MBS/pyrexMD.git
cd pyrexMD
conda install clangxx_osx-64
pip install mpi4py
python -m pip install -e .```
