### 2021-06-29 #3325 @ArtVoro
sorry for the late response.

I did some CI with building and testing the software (took me longer than expected) and reworked the tutorials (juypter notebooks). Now ill expand the online documentation and then revisit the paper.
