### 2021-07-01 #3325 @ArtVoro
Update about the tasks:

- reworked the installation instructions
- added code coverage
- added CI via github actions: build gromacs -> build python package -> run tests -> update documentation
  (build and tests are performed on ubuntu & mac)
- reworked example jupyter notebooks
- added a quick guide to online documentation (https://kit-mbs.github.io/pyrexMD/quick_guide.html)

Todo:

- tasks related to the paper itself
