### 2021-07-01 #3325 @ArtVoro
A few notes:
- The package was private until the JOSS submission so obviously there are no citations about the software. However I am submitting a paper within the next month where I explicitly call the name of the software.
- I intend to upload the package to PyPI but will wait until the end of this review because (as far as I know) it is not possible to reuse a package version number
