### 2021-07-07 #3325 @ArtVoro
@jgostick we pushed a new version of paper.md and extended the citations accordingly. I think all  points mentioned from the reviewers should be covered now.
