### 2021-07-16 #3325 @janash
Thank you @ArtVoro, the documentation is much improved. I would still like to see some additional comments/explanation in the example notebook tutorials.

I have  concerns about code quality and maintainability. Particularly, a large amount of code is repeated when it need not be. Please see the issue I have filed on your repository (https://github.com/KIT-MBS/pyrexMD/issues/1). I would suggest some sort of decorator for the gromacs functions instead of the current implementation.
