### 2021-07-16 #3325 @ArtVoro
thanks for the feedback. the repeated code on gmx.py was actually on my todo list but had low priority. however since its not much effort i can fix it over this weekend.

about the notebooks, i can take another look at it and add more comments/explanation as you suggested but I dont know when exactly. Im working on 2 papers right now and my schedule is already packed. due to the delays in reviews I want to send out the drafts first.
