### 2021-07-17 #3325 @ArtVoro
i fixed the repeated code by calling help functions now.
