### 2021-07-28 #3325 @rosecers
Hey @ArtVoro, I've finally been able to get another look at this and was once again in a tornado of trying to install the package and run the tests.

Comments on installation and testing:
- You have included both a `requirements.txt` and `requirements_python3.6.txt`. You should instead use [environment markers ](https://www.python.org/dev/peps/pep-0508/#environment-markers) to automatically detect the two and avoid confusion.
- In general the installation requirements are quite stringent -- are all of these _exact_ versions necessary?
- Issues with `requirements_python3.6.txt`:
  - `heat=1.0.0` requires `python >=3.7`
  -  ipywidgets 7.4.2 depends on widgetsnbextension~=3.4.0, and the requested version is 3.5.1
- Running `pyrexMD/tests/test_gmx.py` currently results in an error due to a segfault. I am not a regular GROMACS user, so this may be due to my install, but should be addressed nonetheless in case it is due to memory concerns:
``gmx convert-tpr -s pyrexMD/examples/files/protein//../traj/traj_protein.tpr -o /Users/rosecers/work_folders/cosmo/tests/pyrexmd/pyrexMD/traj_protein_protein.tpr``

Comments on what I've been able to get through in the code:
- `gmx.py`: I am concerned with how much of `gmx.py` is modified versions of GromacsWrapper or effectively switch statements. For the functions which are primarily purposed in GromacsWrapper, I would suggest you create a fork of GromacsWrapper to modify the code or propose the modified functions directly to that repository if you think these functions will be appreciated by the community at large. I don't think that they necessitate inclusion into a new package.
- `misc.py`: How much of this is original code, and how much are wrappers to slim down code or add a print statement? This is a rather bloated file, most of which should be deleted.
    - Please delete everything that is effectively

      ``` python
            import module
            def func(*args, **kwargs):
                return module.func(*args, **kwargs)
      ```

      and just refer to the original module functions or a set of module functions instead. There is also code for which the functionality already exists in the current dependencies. Examples: `cwd`, `pwd`, `isdir`, `isfile`, `pathexists`, `realpath`, `dirpath`, `get_filedir`, `mkdir`, and many more.
    - `apply_matplotlib_rc_settings` can just be a matplotlib [style sheet](https://matplotlib.org/stable/tutorials/introductory/customizing.html).
    - is `update_alias_docstring` just for your own use to create docstrings? Where else is it used?
    - What are the functional purposes of `CONFIG`? With the exception `deepcopy_without` and `update_by_alias`, it appears to contain functions that are just renamings, not even wrappers, of built-in functions.

 All in all the documentation, paper, and code have been improved, but still need to be streamlined. I'll try to give another look-over next week.
