### 2021-07-29 #3325 @ArtVoro
Hi @rosecers. Ill adress your points one by one

Comments on installation and testing:
- i did merge both requirements files into one and added the environment markers as you suggested. i expect people to use the setup.py however and just refer to requirements.txt to look up which versions were used as a stable version (but maybe thats just my own thinking).
- i catched some versions which were defined as == but  loosened to >=. The packages which are left to == right now should stay this way as the did break my code at some point (especially nglview)
- about the segfault with gromacs, it must be something with your installation. you can maybe refer to https://github.com/KIT-MBS/pyrexMD/blob/master/.github/workflows/gh-actions-ci.yml -> build-gromacs -> look at used commands for your OS

Comments on code:
- while i see your point with gmx.py I have to disagree with your proposal. The major point is that the current gromacswrapper does exactly what it is supposed to do, i.e. it allows the user to execute gromacs commands via python with the same behavior as it is expected via commandline. The functions which are used in gmx.py are just a very small portion (like 10 out of 105 gromacs functions or so) which have some added functionality because I have to streamline how I handle data in my workflows and are basically just some QoL improvements in my case. However since every gromacs user handles their work slightly different, e.g. by naming their files, data structure etc. forking and then doing pull request seems not reasonable. Besides this way I have a clear overview which functions were used and extended without having 1:1 copy of unchanged code and making it hard to find the differences.
- regarding misc.py however I agree that removing some functions (mainly the os package stuff) was necessary. misc.py is a collection of miscellaneous functions which i use somewhat regular. I wanted to do a cleanup at some point but just forgot about it. I went through the file now and removed all functions which are clearly unnecessary for the package. A few functions which you maybe adressed (mkdir, cd, joinpath) i will keep however since I use them ALOT and they have certain behavior added which I expect and dont want to repeat over and over.
- matplotlib style sheet -> thanks for suggestion. I was not aware this existed.
- update_alias_docstring is just used to by me. I can remove it and just do "alias = function" in the code but I personally dislike that the function name does not get updated there when using "help(function)" or "help(alias)". if you want it removed just tell me.
- CONFIG creates an object which behaves very similar to a dictionary. I use almost everywhere to store and modify config variables in functions or pass some config variables between function calls.  I think you refer to methods such as __setitem__, __delitem__ etc but if I remember correctly such methods are not set by default for an object and i coded them explicitly so that CONFIG  object behaves in a certain way. Besides it helps me to get a quick overview of the allowed kwargs of each function.
