### 2021-08-05 #3325 @ArtVoro
Hi @jgostick, as far as I'm aware I have addressed all tasks which were mentioned following all posts above. Maybe you can give me an example how I could make it more obvious to finalize the review checklists?
The only open tasks might be:

@janash
> Thank you @ArtVoro, the documentation is much improved. I would still like to see some additional comments/explanation in the example notebook tutorials.

Are the comments/explanations of the notebooks sufficient to be accepted? If not can you give me more directions where exactly you want more comments/explanation so i can fix this precisely?

@rosecers
> All in all the documentation, paper, and code have been improved, but still need to be streamlined. I'll try to give another look-over next week.

I'm not quire sure what you mean by streamline so if you can give some clear instructions, I can address this.
