### 2021-08-25 #3325 @rosecers
> @ArtVoro - Thank you for your efforts. I do think the package is much improved from when it was first submitted. However, I can't recommend it for publication at this time.
>
> @ArtVoro @jgostick , please see my comments under "Substantial Scholarly Effort", above.

@janash @jgostick I have similar concerns about the "Substantial Scholarly Effort" and scope and cannot recommend it for publication either.
