### 2021-09-11 #3325 @jgostick
Hi @ArtVoro, at JOSS we do not reject papers, but prefer to allow the authors to work on them until their acceptable. The two reviewers did a very thorough job, but they are out of time to stick around for more rounds of review and review.  Instead, one of JOSS's associate editors (@jarvist) will chime in on some issues that need to be addressed. I think this is largely documentation and explanation of why this code is useful.
