### 2021-09-11 #3325 @jarvist
Dear all,
I'm writing this as an informal editorial assessment. I've read the paper, looked at some of the code, and read the two reviewers comments.

The codes are relatively lightweight wrappers around the GROMACS python bindings, and some analysis tools. There is not much algorithmic complexity, instead these are mostly glue codes. However, what they seem to enable is a Jupyter-notebook based system to construct replica-exchange MD runs, and then analyse the results, all on the computational cluster. This seems quite useful.
I don't have a problem with the coding style, but do point out that thin-wrappers are by their nature a leaky abstraction. Combined with the inter-dependency of the packages used, I expect these codes to be quite brittle to any upstream updates.

I think the paper could be made more clear. I do not find the central statements of "'all-purpose' toolkit" or "workflow-orientated functions" to be specific or descriptive. I would like the PAPER.md to be rewritten to state the functionality that pyrexMD offers more directly.
The figures in the paper are pretty, but they mean little to someone who is not an expert in this particular MD method.

I share some of the concerns of the reviewers about how likely this package will be taken up as a community code. MD research groups generally have internal scripts. I would suggest that simplifying the repository README, and providing teaching examples might help.

The documentation is quite vague: I think it would be useful for everyone if the very first paragraph stated that this package wrapped GROMACS and MDAnalysis python bindings to provide a Jupyter-notebook based environment to design, run and analyse replica-exchange MD.

If the paper and documentation could be made more clear and specific, I would recommend publication.

I am happy to offer more feedback and specifics in this thread, if that would be useful.

Best wishes,
Jarv
