### 2021-09-11 #3325 @ArtVoro
Hi everyone,
first of all: thank you all for the additional feedback.

@jarvist i have a couple of comments and questions regarding your post.

1) did you check the quick guide? i made it to give an initial overview what to expect from the package. Would be nice if you can give feedback regarding this: [ https://kit-mbs.github.io/pyrexMD/quick_guide.html]( https://kit-mbs.github.io/pyrexMD/quick_guide.html).

2) you mention that you want the documentation to be more clear and specific. I asked a few colleagues (all doing either coarse-grained or full atom MD simulations) some time ago to check the API docs and tell me if they find it sufficient to work with. They all told me that given the name of the function, and as I do the naming of the return values and the description of the return values it is generally ok for them. For them this is sufficient information to fully understand what to expect from the functions. Generally speaking, documentation can obviously always be improved but requires lots of work.  At the moment I try aim to hit the "documentation is sufficient" mark.

As it is right now, each module of the API docs states
- what is expected in the module (e.g. HINT: This module contains functions related to (contact-guided) Replica Exchange Molecular Dynamics. It contains mainly functions to automate and speed-up the process of setting up REX simulations.)
- a small 'module example'
- then followed by all functions and their respective doc string (which also sometimes contain a small example)

Therefore I want to know what exactly you expect if you say you want it clearer and more specific. Can you maybe give one or two examples what to improve?

3) regarding the figures: I can obviously only display structure analyses figures and explicitly selected three figures for my target audience (= people doing MD/REMD & Structure Analyses) to display some 'highlights'
- something clear to everyone doing MD (here: RMSD curve with trajectory viewer). I dont think this figure can be improved.
- something related to contact-guided REMD (field-specific, here: TPR curve for ranked bias contacts). The figure shows more than just TPR and therefore requires some additional reading to fully understand but I tryed to explain the additional lines etc.. Should I expand on the figure description or what would you suggest to improve this?
- something related to structure comparison (here: LocalAccuracy figure with GlobalDistanceTest scores). I guess the only way to improve this figure is to explain shortly what GlobalDistanceTest means in case somebody does not know it yet and intoduce the two score variants and abbreviations TS (total score) and HA (high accuracy).

I hope with your answers I can understand better what you expect so that I can improve those aspects.
