### 2021-09-11 #3325 @ArtVoro
quick note on current todos:

> I think it would be useful for everyone if the very first paragraph stated that this package wrapped GROMACS and MDAnalysis python bindings to provide a Jupyter-notebook based environment to design, run and analyse replica-exchange MD.

easy fix, will do that tomorrow

- I will sit down and rework PAPER.md and try to explain it clearer what exactly the pyrexMD does and offers, as suggested by @jarvist
