### 2021-09-14 #3325 @jarvist
Dear ArtVoro,

I think the detailed documentation (APIs & docstrings) is of a fine standard for publication.

What I feel is either missing or too vague are the higher level descriptions, to understand what the package does, what it enables, why should someone use it.

I did see your quickstart guide, but again I didn't really understand the motivation for each section. The 'normal MD' looked like a standard command-line GROMACS workflow. But I look at those commands and wonder: do I have to supply the mdp files? What are the inputs and outputs? Do the functions generate files as side effects, and is this how the whole process hangs together?
Also, ideally those examples should be runnable. (So provide a demo .pdb file etc.)
Does your first example contain a typo? It looks like a mix of python and command line. Having them as runnable iPython notebooks helps avoid such errors.
> gmx.grompp(-f"md.mdp", o="traj.tpr", c="npt.gro", t="npt.cpt")

A showing a side-by-side comparison between your package and doing replica-exchange 'manually' much also be illuminating for any potential users.

Similarly with the figures - I can see them, I can see they look quite nice, but I don't understand what the motivation for them is. Does the package provide a particularly easy way of generating such figures? Is doing this analysis in pyREX / Jupyter notebooks so much more easy?

I'm certainly not saying that the documentation needs totally rewriting, nor massively expanding - but more that it needs more sign posting.

Best wishes,

Jarv
