### 2021-09-14 #3325 @ArtVoro
understood. i will try to improve those aspects today and tell you when both documentation and paper are reworked.

> Also, ideally those examples should be runnable. (So provide a demo .pdb file etc.)
there is already an example folder with jupyter notebooks in the package included (with all necessary files). the readme page also states how to run them.
