### 2021-09-17 #3325 @ArtVoro
Hello @jarvist,

I collected additional feedback and updated the paper, readme and quick guide sections of the github documentation.

- git repo: [https://github.com/KIT-MBS/pyrexMD](https://github.com/KIT-MBS/pyrexMD)
- readme: [https://kit-mbs.github.io/pyrexMD/about_pyrexMD.html](https://kit-mbs.github.io/pyrexMD/about_pyrexMD.html)
- quick guide: [https://kit-mbs.github.io/pyrexMD/quick_guide.html](https://kit-mbs.github.io/pyrexMD/quick_guide.html)

The biggest changes are usually at the very beginning

- paper and readme: I now explicitly state who should be interested in this package and what its main goals are.
- quick guide: reworked top section, module overview section added

Additionally, I made many small changes in the paper and quick guide with the focus of being more specific.
I also noted that the code snippets are taken from the example folder and can be run locally to test the behavior.

Best regards
Arthur
