### 2021-09-17 #3325 @ArtVoro
@jarvist one short comment on your previous statement:
> I don't have a problem with the coding style, but do point out that thin-wrappers are by their nature a leaky abstraction. Combined with the inter-dependency of the packages used, I expect these codes to be quite brittle to any upstream updates.

I personally dont expect the code to be brittle. My main dependencies for wrapping are given by Gromacs and MDAnalysis.
Gromacs is very stable and didnt have any big changes since their command-line syntax change in (i believe)  version 2016. Besides I only use a small portion of very basic gromacs commands to streamline the setups.

Regarding MDAnalysis, I expect problems only during major version updates if they change some data structures. In this case they just recently had a major version update from 1.1.1 to 2.0.0 in August. 2 functions of my code got affected but I already addressed all changes in August.
