### 2021-10-05 #3325 @ArtVoro
Hi @jarvist ,

I just wanted to check on the current status of the review. Seems like you didn't find the time yet to check the changes. I hope that we can finalize the review soon because I'm about to submit another paper where i want to reference the pyrexMD paper.

Best regards,
Arthur
