### 2021-11-08 #3325 @ArtVoro
Hi @jarvist, do you have any feedback regarding the corrections? almost 2 months have passed since the last update. I think I adressed all your points but am waiting for response now.

Best regards,
Arthur
