### 2021-11-30 #3325 @jarvist
Dear @ArtVoro ,
Thank you for your patience and apologies for my very late 2nd review! Teaching commitments & sick children have made for a very busy Autumn so far.

I've had a read through your paper, as well as the README.md and landing page of the documentation. I think the text is much improved, and it clearly states what the package does and why someone should be interested in it. As well as making a better publication, I'm think this will increase the likelihood of someone using your code.

I think the captions for the figures are useful for a peer in a nearby field to understand more of the capabilities. The text around the 'examples' section flows much better, and similarly motivates and explains what is going on. I think abstracting some of this text for the paper is a good idea.

Overall the documentation and paper have been made more clear and more specific, and I think they are now of sufficient quality to recommend publication.

Very minor typo - GROMACS is usually in all caps (as its an acronym), two instances were written as 'Gromacs' in your paper. There may be other instances in your main documentation.
