### 2021-11-30 #3325 @ArtVoro
@jarvist
thank you very much for your kind words and your help with the review.

You are right with the all caps of GROMACS. I checked all files and fixed all instances in the paper as well as some within the documentation. The updated version is already pushed to the repo.

@jgostick
what are the next steps?
