### 2021-11-30 #3325 @ArtVoro
@jgostick

>     * [x]  Double check authors and affiliations (including ORCIDs)
>
>     * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

1.0

>     * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.

10.5281/zenodo.5744760

>     * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
