### 2021-11-30 #3325 @ArtVoro
@jgostick thank you very much for your assistance throughout the review.

The proof seems fine. Do we have to use the "@wheadon accept deposit=true" command or is this for the editor in chief?
