### 2021-11-30 #3325 @kthyng
@janash, @rosecers I am not sure if you've followed the developments on this JOSS submission after you worked on reviews. In short, we have had another person contribute to the review and it has been found in good enough shape for acceptance. Currently you are both listed as reviewers and we'd like to list you as reviewers in recognition of your efforts here, but understand if you do not wish to be listed. Could you please comment here as to your preference either way?
