### 2021-12-01 #3325 @kthyng
@ArtVoro things look good, though please preserve capitalization for Jupyter in a reference with {} in the .bib file around the word or character. Also check the other references carefully for correct capitalization.
