### 2021-12-01 #3325 @ArtVoro
@kthyng i checked the reference titles based on the proof from [https://github.com/openjournals/joss-papers/pull/2780](https://github.com/openjournals/joss-papers/pull/2780). They should all be correct now.
