### 2021-12-01 #3325 @kthyng
Congrats on your new publication @ArtVoro! Many thanks to editor @jgostick and reviewers @janash, @rosecers, and @jarvist on your time, hard work, and expertise!!
